requirements.txt
asgiref
Django
django-environ
django-filter
djangorestframework
Markdown
pytz
sqlparse